# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API APP",
        "description": "API Documentation",
        "contact": {
            "email": "chiranad@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "Student Register",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "successfully register"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "payload": {
                                            "properties": {
                                                "id": {
                                                    "type": "int",
                                                    "example": "1"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "chiran"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "chiran1@gmail.com"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2022-01-13T18:16:24.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2022-01-13T18:16:24.000000Z"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "12|wA5KIi8F8e6BBdW79NWw1iPXJ3Xy4YeaArlc84r7"
                                                }
                                            },
                                            "type": "object",
                                            "format": "query"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Login",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "successfully login"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "payload": {
                                            "properties": {
                                                "id": {
                                                    "type": "int",
                                                    "example": "1"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "chiran"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "chiran1@gmail.com"
                                                },
                                                "email_verified_at": {
                                                    "type": "string",
                                                    "example": "null"
                                                },
                                                "status": {
                                                    "type": "int",
                                                    "example": "1"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2022-01-13T18:16:24.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2022-01-13T18:16:24.000000Z"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "12|wA5KIi8F8e6BBdW79NWw1iPXJ3Xy4YeaArlc84r7"
                                                }
                                            },
                                            "type": "object",
                                            "format": "query"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}